requirements.txt
django
djangorestframework
django-request-logging
psycopg2-binary
